requirements.txt
# Web Scraping
requests
beautifulsoup4
lxml

# Data Processing
pandas
python-dotenv

# RAG Pipeline - Embeddings & Vector Store
sentence-transformers
chromadb
langchain
langchain-community
langchain-openai

# LLM Integration
openai

# Utilities
tqdm

# Web Interface
streamlit
